+++ soxspipe/utKit.py
@@ -47,7 +47,7 @@ class utKit(utKit):
                 formatter: console_style
                 stream: ext://sys.stdout
         root:
-            level: DEBUG
+            level: WARNING
             handlers: [console]"""
 
         # Override Variable Data Atrributes